Add ability to disable parallel-run tests in hugolib
Add disableParallel boolean in testhelpers_test.go; if set to
true, no tests in hugolib will be run in parallel. This makes it
a little easier to track down weird test issues. All the
t.Parallel() calls were changed to t_Parallel(t) calls in order
to make this work.