Add --workers, update num workers defaults
Adds --workers=NUM to the hugo and hugo server commands,
to allow the user to manually set the number of goroutines
spawned. This is useful if the defaults aren't optimum for
specific systems. It's also useful for debugging; setting
--workers=1 means log messages are serialized. This overrides
defaults in renderPages, newSiteContentProcessor, and
newCapturer.

The GOMAXPROCS environment variable is unused on most systems,
and not needed with --workers, so stop using it.

Changes renderPages so it uses 2 * runtime.NumCPU workers,
with a minimum of 4. Previously this was essentially a hard-coded
4, since GOMAXPROCS was rarely defined.

Changes newSiteContentProcessor to use 3 * runtime.NumCPU workers,
with minimum 6, under the reasonable belief that more disk I/O is
happening here than in renderPages.

TBD - the current capturer code starts a goroutine for each nested
dir, to the limit of numWorkers. This might not actually be a
good idea, as directory scanning isn't necessarily very parallel
(depends on the filesystem). Look for best practices in other
systems.