Unify markup-created links with template links
This fixes bad links generated with combinations of baseURL,
canonifyURLs, and relativeURLs, removing the need for workarounds
in templates and config. For all settings now, links are
correct, without doubled-up baseURL paths.

Links in markup-rendered output now match the links made with
template actions. There were hacks in the Hugo code to try to
make things work out, and those are now removed (e.g. GetBasePath
doesn't need to check for canonify, etc).

GetBasePath is now just an accessor to the fixed-up version of
BaseURL.Path(), and no longer needs an isRelative bool parameter.

PrependBasePath now no longer needs an isAbs bool parameter.

RenderBytes does a post-fixup on links in markup-rendered HTML,
so that they respect BasePath. This does add extra time to
rendering a site, but it is unavoidable if you want working links
given the current config options allowed; markup-processed links
have to match how template actions work. The other option would
be to check template actions to not insert BasePath and then fix
it all up in publish, but that is a much more invasive task, and
not actually completely possible with the information available
to template actions at the moment (there are complications with
non-HTML/non-XML output formats).

The relativeURLs path in renderAndWritePage/XML is a little
awkward now, but the whole relative path generation bit needs
to be fixed anyway - it makes unwieldly links and they don't
make the site truly relocatable.