Move AdjustMarkupLinks to urlreplacers
The helpers package is too low in the hierarchy to access
the transform packages, so move AdjustMarkupLInks to a better
home in transform/urlreplacers, and rename it to something
more appropriate, ConvertSiteToPathAbs. Since we can't hook
RenderBytes, we add the appropriate calls to each RenderBytes
site. Maybe make a new wrapper function for this? There's no
obvious place for it, though.

The relativeURLs functionality is broken, and the fix here is
to do it for real, not just trying to prefix some .. path.