--verbose and --log take an optional level value
This updates --verbose and --log to take an optional log level
value (--verbose for console logging, --log for file logging):

  --verbose=LEVEL
  --log=LEVEL

The allowed levels are:

  --verbose=info   --log=info
  --verbose=warn   --log=warn
  --verbose=debug  --log=debug
  --verbose=trace  --log=trace

These set the corresponding log level for console or file log,
and override existing vars like "debug" or "verboseLog". As
a courtesy, the existing verbose flag is still a boolean and
is true if --verbose is used in any form.