Process markup to path-absolute-URL links
Hugo sites need path-absolute-links (with the path from baseURL) in
their HTML, but markup doesn't know about that, and makes
site-absolute-URL links. There was an attempt to kludge around this,
but it only works for a subset of cases; there were many cases where
bad URLs were generated.

This patch adds a post-process after markup generation to turn any
site-absolute-URL links into path-absolute-URL links. This matches
how Hugo templates work - for template generation, we insert either
path-relative-URL, path-absolute-URL, or absolute-URL links. So, we
make sure that markup-generated HTML has only path-relative-URL,
path-absolute-URL, or absolute-URL links.

For all settings now, generated links are correct, without doubled-up
baseURL paths or site-absolute-URLs where we wanted path-absolute-URLs.

The hacks around canonifyURLs in GetBasePath, PrependBasePath, RelURL,
getMenusFromConfig, and tests are no longer needed, and are removed.
This makes the code a little less fragile. In particular, a number of
tests had canonifyURLs hacks embedded in them.

GetBasePath is now just an accessor to the fixed-up version of
BaseURL.Path(), and no longer needs an isRelative bool parameter.

RelURL, in the process of removing the canonify hack, was completely
rewritten to be a lot clearer (its previous code had gotten a little
too complex and fragile to be easily understandable).

PrependBasePath now no longer needs an isAbs bool parameter. A very
similar routine AddContextRoot has been removed, since nothing was
calling it any more.

RenderBytes does a post-fixup on links in markup-rendered HTML, so
that they respect BasePath. This does add extra time to rendering a
site, but it is unavoidable if you want working links given the
current config options allowed; markup-processed links have to match
how template actions work. The other option would be to check template
actions to not insert BasePath and then fix it all up in publish, but
that is a much more invasive task, and not actually completely possible
with the information available to template actions at the moment
(there are complications with non-HTML/non-XML output formats). The
other other option would be to switch to a different Markdown
renderer with more control over the generated HTML.